refactor: remove query client from example
