+++ example/package.json
@@ -22,7 +22,7 @@
     "@types/react": "link:../node_modules/@types/react",
     "@types/react-dom": "link:../node_modules/@types/react-dom",
     "immutable": "4.0.0-rc.12",
-    "react": "^17.0.1",
+    "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-scripts": "3.4.3",
     "typescript": "link:../node_modules/typescript"
