+++ src/utils/utils.js
@@ -1,0 +1,21 @@
+// todo: use graasp-utils repo
+export const getParentsIdsFromPath = (path, { ignoreSelf = false } = {}) => {
+  if (!path) {
+    return [];
+  }
+
+  let p = path;
+  // ignore self item in path
+  if (ignoreSelf) {
+    // split path in half parents / self
+    // eslint-disable-next-line no-useless-escape
+    const els = path.split(/\.[^\.]*$/);
+    // if els has only one element, the item has no parent
+    if (els.length <= 1) {
+      return [];
+    }
+    [p] = els;
+  }
+  const ids = p.replace(/_/g, '-').split('.');
+  return ids;
+};
